justfit-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.7.2
pydantic==2.10.4
python-dotenv==1.0.0
firebase-admin==6.3.0
python-multipart==0.0.6